requirements.txt
aiogram==3.3.0
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
aiohttp
requests